refactor(renderer): extract tray icon logic to custom hook
